polyline-64bit-compatibility-fix
Fixes a problem with the polyline function on a 64-bit platform.
The wxDC_DrawLines function now properly indexes into the array
with points.

darcs-hash:20070313080116-fafd9-27751f0f90b85e7190a1dc146c37680990c42fa8.gz

+++ wxc/src/ewxw/eljdc.cpp
@@ -64,7 +64,7 @@ EWXWEXPORT(void, wxDC_DrawLines)(void* _obj, int n, void* x, void* y, int xoffse
 	wxPoint* lst = (wxPoint*)malloc (n * sizeof(wxPoint));
 	
 	for (int i = 0; i < n; i++)
-		lst[i] = wxPoint(((int*)x)[i], ((int*)y)[i]);
+		lst[i] = wxPoint((int)((intptr_t*)x)[i], (int)((intptr_t*)y)[i]);
 	
 	((wxDC*)_obj)->DrawLines(n, lst, (wxCoord)xoffset, (wxCoord)yoffset);
